Speed up hr_locoh() by using st_union() in place of rgeos::gBuffer()
For a track object with 20,000 rows, this edit speeds up hr_locoh()
nearly three-fold, cutting processing time from 35 to 12.7 seconds on
my laptop. On my computer, a track object with 30,000 rows crashes R
with the previous code, but runs to completion in 20.6 seconds after
this edit.

The one caveat is that the area of the MULTIPOLYGON outputted by this
code is slightly different (on the order of 0.03%) than that of the
previous output, due to its replacement of a call to
`rgeos::gBuffer()` with a call to `st_union(..., by_feature=TRUE)`.

There is a note in the block of code that I've edited that reads
"buffer is necessary, to overcome some topology errors if the polygon
is quasi a line". Before accepting this commit, it might be good to
test with such a potentially problematic polygon to see whether it
causes issues in the results returned by `st_union()`.