Speed up computation of proportion of points seen hr_locoh
Using findInterval(), I've vectorized and sped up the "for" loop that
was marked with the comment "this is still slow". The new code
computed the same value for vector `pp` as the for loop that it
replaces (as expected) for the ~20 cases I've tested it on. For a
track object with 20,000 rows, this commit cuts total computation time
for hr_locoh.track_xy() by ~35%, from 55 seconds to 34 seconds on my
laptop.

(One thing to note: like the for loop it replaces, this algorithm
doesn't ever include the central point of each knn polygon in the set
of points "seen" in that polygon. I'm not sure if that's intentional,
but in any case my algorithm faithfully reproduces the existing behavior.)