Fix hr_locoh to reflect fact that hulls contain their "focal" point
In `hr_locoh()`, `zz` is a list, each of whose elements is a
`SpatialPointsDataFrame` holding the points used to construct the hull
around a given point. `zz` is also used (after being reordered to
create `ff`) to compute the cumulative proportion of all points
included within the home range by the addition to it of the nth
smallest hull. For both purposes, it's important that each element of
zz include not only the nearby points identified by the algorithm
specified by `type=` and by the user's choice of `n=`, but also the
point with respect to which the other points were identified, as it
should always be part of the hull.

The simple tweak proposed here adds the focal point to each element of
`zz`. FWIW, across a large number of cases I've examined, this change
makes the isopleths returned by `amt::hr_locoh()` match much more
closely (or completely) those produced by `tlocoh::lhs.iso.add()`,
whose author (Andy Lyons) was also one of the coauthors on Getz et
al. (2007), the paper that introduced LoCoH.